Update PDE.py
roll back dtype  xInt.append(np.linspace(x[i][0], x[i][1], pointCount, dtype=np.float64))

+++ PDEs/PDE.py
@@ -324,8 +324,8 @@ class PDENeuralNetwork:
         xInt = []
         for i in range(len(x)):
             if isinstance(x[i], tuple):
-                xInt.append(np.linspace(x[i][0], x[i][1], pointCount))
-#                 xInt.append(np.linspace(x[i][0], x[i][1], pointCount, dtype=np.float64))
+#                 xInt.append(np.linspace(x[i][0], x[i][1], pointCount))
+                xInt.append(np.linspace(x[i][0], x[i][1], pointCount, dtype=np.float64))
                     
 
             else:
